Bump mask to be 64 bits - this fixes seeing MCS rates > 19.
Tested:

* AR9380, STA mode, associated to a 3x3 AP (also FreeBSD)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@280361 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ tools/tools/ath/athratestats/main.c
@@ -115,7 +115,7 @@ static void
 ath_sample_stats(struct ath_ratestats *r, struct ath_rateioctl_rt *rt,
     struct sample_node *sn)
 {
-	uint32_t mask;
+	uint64_t mask;
 	int rix, y;
 
 	PRINTMSG("static_rix (%d) ratemask 0x%llx\n",
